+++ dca/tl_news_archive.php
@@ -31,6 +31,35 @@ array_insert($GLOBALS['TL_DCA']['tl_news_archive']['list']['global_operations'],
     )
 ));
 
+/**
+ * Add palettes to tl_news_archive
+ */
+$GLOBALS['TL_DCA']['tl_news_archive']['palettes']['__selector__'][] = 'limitCategories';
+$GLOBALS['TL_DCA']['tl_news_archive']['palettes']['default'] = str_replace('jumpTo;', 'jumpTo;{categories_legend},limitCategories;', $GLOBALS['TL_DCA']['tl_news_archive']['palettes']['default']);
+$GLOBALS['TL_DCA']['tl_news_archive']['subpalettes']['limitCategories'] = 'categories';
+
+/**
+ * Add fields to tl_news_archive
+ */
+$GLOBALS['TL_DCA']['tl_news_archive']['fields']['limitCategories'] = array
+(
+    'label'                   => &$GLOBALS['TL_LANG']['tl_news_archive']['limitCategories'],
+    'exclude'                 => true,
+    'inputType'               => 'checkbox',
+    'eval'                    => array('submitOnChange'=>true),
+    'sql'                     => "char(1) NOT NULL default ''"
+);
+
+$GLOBALS['TL_DCA']['tl_news_archive']['fields']['categories'] = array
+(
+    'label'                   => &$GLOBALS['TL_LANG']['tl_news_archive']['categories'],
+    'exclude'                 => true,
+    'inputType'               => 'treePicker',
+    'foreignKey'              => 'tl_news_category.title',
+    'eval'                    => array('multiple'=>true, 'fieldType'=>'checkbox', 'foreignTable'=>'tl_news_category', 'titleField'=>'title', 'searchField'=>'title', 'managerHref'=>'do=news&table=tl_news_category'),
+    'sql'                     => "blob NULL"
+);
+
 class tl_news_archive_categories extends Backend
 {