+++ languages/pl/tl_news_archive.php
@@ -16,3 +16,14 @@
  * Global operations
  */
 $GLOBALS['TL_LANG']['tl_news_archive']['categories'] = array('Kategorie', 'Dodawaj i zarządaj kategoriami aktualności.');
+
+/**
+ * Legends
+ */
+$GLOBALS['TL_LANG']['tl_news_archive']['categories_legend'] = 'Ustawienia kategorii';
+
+/**
+ * Fields
+ */
+$GLOBALS['TL_LANG']['tl_news_archive']['limitCategories'] = array('Ogranicz kategorie', 'Ogranicz dostępne kategorie dla tego archiwum aktualności.');
+$GLOBALS['TL_LANG']['tl_news_archive']['categories']      = array('Kategorie aktualności', 'Wybierz jedną lub więcej kategorii aktualności.');